Update publications page
So that it actually displays a list of publications.  It's not a complete list yet, but it's definitely a start.

I do, however, need to add some conditional logic to my publications shortcode, so that some things are twiddled differently depending on the circumstances of the publication.  E.g. those without a DOI shouldn't have a doi line shown.